+++ src/modules/cards/infrastructure/http/routes/index.ts
@@ -27,6 +27,8 @@ import { SearchCollectionsController } from '../controllers/SearchCollectionsCon
 import { GetOpenCollectionsWithContributorController } from '../controllers/GetOpenCollectionsWithContributorController';
 import { GetCollectionPageByAtUriController } from '../controllers/GetCollectionPageByAtUriController';
 import { GetCollectionsForUrlController } from '../controllers/GetCollectionsForUrlController';
+import { GetRecommendedCollectionsForUrlController } from '../controllers/GetRecommendedCollectionsForUrlController';
+import { GetRecommendedOpenCollectionsForUrlController } from '../controllers/GetRecommendedOpenCollectionsForUrlController';
 import { GetCollectionFollowersController } from '../controllers/GetCollectionFollowersController';
 import { GetCollectionFollowersCountController } from '../controllers/GetCollectionFollowersCountController';
 import { GetCollectionContributorsController } from '../controllers/GetCollectionContributorsController';
@@ -61,6 +63,8 @@ export function registerCardsModuleRoutes(
   getMyCollectionsController: GetMyCollectionsController,
   getCollectionsController: GetUserCollectionsController,
   getCollectionsForUrlController: GetCollectionsForUrlController,
+  getRecommendedCollectionsForUrlController: GetRecommendedCollectionsForUrlController,
+  getRecommendedOpenCollectionsForUrlController: GetRecommendedOpenCollectionsForUrlController,
   searchCollectionsController: SearchCollectionsController,
   getOpenCollectionsWithContributorController: GetOpenCollectionsWithContributorController,
   getCollectionFollowersController: GetCollectionFollowersController,
@@ -99,6 +103,8 @@ export function registerCardsModuleRoutes(
     getCollectionsController,
     getCollectionPageByAtUriController,
     getCollectionsForUrlController,
+    getRecommendedCollectionsForUrlController,
+    getRecommendedOpenCollectionsForUrlController,
     searchCollectionsController,
     getOpenCollectionsWithContributorController,
     getCollectionFollowersController,