+++ src/shared/infrastructure/http/factories/ControllerFactory.ts
@@ -21,6 +21,7 @@ import { GetMyCollectionsController } from '../../../../modules/cards/infrastruc
 import { GetGlobalFeedController } from '../../../../modules/feeds/infrastructure/http/controllers/GetGlobalFeedController';
 import { GetGemActivityFeedController } from '../../../../modules/feeds/infrastructure/http/controllers/GetGemActivityFeedController';
 import { GetFollowingFeedController } from '../../../../modules/feeds/infrastructure/http/controllers/GetFollowingFeedController';
+import { GetBskyFollowingFeedController } from '../../../../modules/feeds/infrastructure/http/controllers/GetBskyFollowingFeedController';
 import { GetSimilarUrlsForUrlController } from '../../../../modules/search/infrastructure/http/controllers/GetSimilarUrlsForUrlController';
 import { SemanticSearchUrlsController } from '../../../../modules/search/infrastructure/http/controllers/SemanticSearchUrlsController';
 import { RecommendedCardsController } from '../../../../modules/search/infrastructure/http/controllers/RecommendedCardsController';
@@ -48,6 +49,8 @@ import { GetOpenCollectionsWithContributorController } from 'src/modules/cards/i
 import { GetCollectionPageByAtUriController } from 'src/modules/cards/infrastructure/http/controllers/GetCollectionPageByAtUriController';
 import { GetUrlStatusForMyLibraryController } from '../../../../modules/cards/infrastructure/http/controllers/GetUrlStatusForMyLibraryController';
 import { GetLibrariesForUrlController } from '../../../../modules/cards/infrastructure/http/controllers/GetLibrariesForUrlController';
+import { GetRecommendedCollectionsForUrlController } from '../../../../modules/cards/infrastructure/http/controllers/GetRecommendedCollectionsForUrlController';
+import { GetRecommendedOpenCollectionsForUrlController } from '../../../../modules/cards/infrastructure/http/controllers/GetRecommendedOpenCollectionsForUrlController';
 import { GetCollectionsForUrlController } from '../../../../modules/cards/infrastructure/http/controllers/GetCollectionsForUrlController';
 import { GetNoteCardsForUrlController } from '../../../../modules/cards/infrastructure/http/controllers/GetNoteCardsForUrlController';
 import { GetMyNotificationsController } from '../../../../modules/notifications/infrastructure/http/controllers/GetMyNotificationsController';
@@ -141,6 +144,8 @@ export interface Controllers {
   getUrlStatusForMyLibraryController: GetUrlStatusForMyLibraryController;
   getLibrariesForUrlController: GetLibrariesForUrlController;
   getCollectionsForUrlController: GetCollectionsForUrlController;
+  getRecommendedCollectionsForUrlController: GetRecommendedCollectionsForUrlController;
+  getRecommendedOpenCollectionsForUrlController: GetRecommendedOpenCollectionsForUrlController;
   getNoteCardsForUrlController: GetNoteCardsForUrlController;
   getCollectionFollowersController: GetCollectionFollowersController;
   getCollectionFollowersCountController: GetCollectionFollowersCountController;
@@ -161,6 +166,7 @@ export interface Controllers {
   getGlobalFeedController: GetGlobalFeedController;
   getGemActivityFeedController: GetGemActivityFeedController;
   getFollowingFeedController: GetFollowingFeedController;
+  getBskyFollowingFeedController: GetBskyFollowingFeedController;
   // Search controllers
   getSimilarUrlsForUrlController: GetSimilarUrlsForUrlController;
   semanticSearchUrlsController: SemanticSearchUrlsController;
@@ -381,6 +387,14 @@ export class ControllerFactory {
       getCollectionsForUrlController: new GetCollectionsForUrlController(
         useCases.getCollectionsForUrlUseCase,
       ),
+      getRecommendedCollectionsForUrlController:
+        new GetRecommendedCollectionsForUrlController(
+          useCases.getRecommendedCollectionsForUrlUseCase,
+        ),
+      getRecommendedOpenCollectionsForUrlController:
+        new GetRecommendedOpenCollectionsForUrlController(
+          useCases.getRecommendedOpenCollectionsForUrlUseCase,
+        ),
       getNoteCardsForUrlController: new GetNoteCardsForUrlController(
         useCases.getNoteCardsForUrlUseCase,
       ),
@@ -438,6 +452,9 @@ export class ControllerFactory {
       ),
       getFollowingFeedController: new GetFollowingFeedController(
         useCases.getFollowingFeedUseCase,
+      ),
+      getBskyFollowingFeedController: new GetBskyFollowingFeedController(
+        useCases.getBskyFollowingFeedUseCase,
       ),
       // Search controllers
       getSimilarUrlsForUrlController: new GetSimilarUrlsForUrlController(