+++ src/types/src/api/index.ts
@@ -28,6 +28,7 @@ export * from './collections/getForUrl';
 export * from './collections/listContributed';
 export * from './collections/getFollowers';
 export * from './collections/getContributors';
+export * from './collections/recommendedForUrl';
 
 export * from './users/getProfile';
 export * from './users/login';
@@ -42,6 +43,7 @@ export * from './users/onboardingState';
 export * from './feeds/global';
 export * from './feeds/gem';
 export * from './feeds/following';
+export * from './feeds/bskyFollowing';
 
 export * from './notifications/list';
 export * from './notifications/unreadCount';