+++ src/types/src/routes.ts
@@ -71,6 +71,10 @@ export const paths = {
   collectionFollowers: '/network.cosmik.collection.getFollowers',
   collectionFollowersCount: '/network.cosmik.collection.getFollowerCount',
   collectionContributors: '/network.cosmik.collection.getContributors',
+  recommendedCollectionsForUrl:
+    '/network.cosmik.collection.getRecommendedForUrl',
+  recommendedOpenCollectionsForUrl:
+    '/network.cosmik.collection.getRecommendedOpenForUrl',
   // users / actor
   myProfile: '/network.cosmik.actor.getMyProfile',
   userProfile: '/network.cosmik.actor.getProfile',
@@ -102,6 +106,7 @@ export const paths = {
   globalFeed: '/network.cosmik.feed.getGlobal',
   gemFeed: '/network.cosmik.feed.getGem',
   followingFeed: '/network.cosmik.feed.getFollowing',
+  bskyFollowingFeed: '/network.cosmik.feed.getBskyFollowing',
   // notifications
   myNotifications: '/network.cosmik.notification.list',
   unreadCount: '/network.cosmik.notification.getUnreadCount',
@@ -170,6 +175,16 @@ export const routes = {
     followers: defineRoute(paths.collectionFollowers, 'GET', false),
     followersCount: defineRoute(paths.collectionFollowersCount, 'GET', false),
     contributors: defineRoute(paths.collectionContributors, 'GET', false),
+    recommendedForUrl: defineRoute(
+      paths.recommendedCollectionsForUrl,
+      'GET',
+      true,
+    ),
+    recommendedOpenForUrl: defineRoute(
+      paths.recommendedOpenCollectionsForUrl,
+      'GET',
+      true,
+    ),
   },
   users: {
     myProfile: defineRoute(paths.myProfile, 'GET', true),
@@ -201,6 +216,7 @@ export const routes = {
     global: defineRoute(paths.globalFeed, 'GET', false),
     gem: defineRoute(paths.gemFeed, 'GET', false),
     following: defineRoute(paths.followingFeed, 'GET', true),
+    bskyFollowing: defineRoute(paths.bskyFollowingFeed, 'GET', true),
   },
   notifications: {
     myNotifications: defineRoute(paths.myNotifications, 'GET', true),